requirements.txt
streamlit>=1.32.0
pillow>=10.0.0
pytesseract>=0.3.10
rapidfuzz>=3.6.1
openai>=1.12.0
ollama>=0.1.7
python-dotenv>=1.0.0
